requirements.txt
# Core dependencies
tensorflow>=2.8.0
magenta>=2.1.4
numpy>=1.21.0
flask>=2.0.0
flask-cors>=3.0.10
requests>=2.25.0
pretty_midi>=0.2.9
librosa==0.7.2

# Cloud and deployment
google-cloud-compute>=1.0.0
google-cloud-storage>=2.0.0
gunicorn>=20.1.0

# MIDI processing
mido>=1.2.0
python-rtmidi>=1.4.0

# Audio processing
soundfile>=0.10.0
scipy>=1.7.0

# Utilities
python-dotenv>=0.19.0
click>=8.0.0
tqdm>=4.62.0
colorama>=0.4.4

# Development
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0
